Display jobs on the map as pins
