+++ Machbarschaft/VCs/MainViewController.swift
@@ -8,26 +8,37 @@
 
 import UIKit
 import OverlayContainer
+import CoreLocation
 
 class MainViewController: UIViewController {
     
     @IBOutlet weak var overlayContainerView: UIView!
     let overlayController = OverlayContainerViewController(style: .flexibleHeight)
     let mapViewController = viewController(withID: "MapViewController") as? MapViewController
+    let jobTableViewController = viewController(withID: "JobMenuViewController") as? JobMenuViewController
+    
+    var locationManager: CLLocationManager!
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        guard let menuVC = viewController(withID: "JobMenuViewController") as? JobMenuViewController else { return }
         overlayController.delegate = self
-        overlayController.viewControllers = [menuVC]
+        overlayController.viewControllers = [jobTableViewController!]
         addChild(overlayController, in: view)
         if mapViewController != nil {
             addChild(mapViewController!, in: overlayContainerView)
         }
         
-        menuVC.shouldSegueToJobSummary = { job in
+        jobTableViewController!.shouldSegueToJobSummary = { job in
             self.performSegue(withIdentifier: "Main_to_JobSummary", sender: job)
+        }
+        
+        if (CLLocationManager.locationServicesEnabled()) {
+            locationManager = CLLocationManager()
+            locationManager.delegate = self
+            locationManager.desiredAccuracy = kCLLocationAccuracyBest
+            locationManager.requestWhenInUseAuthorization()
+            locationManager.startUpdatingLocation()
         }
     }
     
@@ -59,5 +70,20 @@ extension MainViewController: OverlayContainerViewControllerDelegate {
     
     func numberOfNotches(in containerViewController: OverlayContainerViewController) -> Int {
         OverlayNotch.allCases.count
+    }
+}
+
+
+extension MainViewController: CLLocationManagerDelegate {
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        
+        if let location = locations.last {
+            let api = API
+            api?.loadJobs(location: location.coordinate, completion: { (jobs) in
+                self.mapViewController?.update(userLocation: location.coordinate, jobs: jobs)
+                self.jobTableViewController?.update(jobs: jobs)
+            })
+        }
     }
 }
